@@ -5,12 +5,14 @@ USAGE:
     $ idpconfgen fastaext [PDBS]
 """
 import argparse
-import itertools as it
 from multiprocessing import Manager
 
 from idpconfgen import Path, log
-from idpconfgen.libs import libcli, libio, libpdb, libstructure, libmulticore
+from idpconfgen.libs import libcli
 from idpconfgen.logger import S, T, init_files
+from idpconfgen.libs.libhigherlevel import get_fastas
+from idpconfgen.libs.libmulticore import pool_function
+from idpconfgen.libs.libio import FileReaderIterator, save_dictionary
 
 
 LOGFILESNAME = 'idpconfgen_fastaext'
@@ -26,45 +28,28 @@ ap = libcli.CustomParser(
     usage=_us,
     formatter_class=argparse.RawDescriptionHelpFormatter,
     )
-# https://stackoverflow.com/questions/24180527
 
-ap.add_argument(
-    'pdbs',
-    help='PDB file list.',
-    nargs='+',
-    )
+libcli.add_argument_pdb_files(ap)
 
 ap.add_argument(
     '-o',
     '--output',
-    help=( "The output file containing the PDBID and "
+    help=(
+        "The output file containing the PDBID and "
         "respective FASTA sequence information. "
         "Defaults to sys.stdout (prints to console)."
         ),
     type=Path,
     default=None,
+    const='fasta.json',
+    action=libcli.CheckExt({'.json'}),
     )
 
 libcli.add_argument_ncores(ap)
 
 
-def _load_args():
-    cmd = ap.parse_args()
-    return cmd
-
-
-def maincli():
-    """
-    Execute main client function.
-
-    Reads command line arguments and executes logic.
-    """
-    cmd = _load_args()
-    main(**vars(cmd))
-
-
 def main(
-        pdbs,
+        pdb_files,
         output=None,
         ncores=1,
         **kwargs,
@@ -85,38 +70,22 @@ def main(
     init_files(log, LOGFILESNAME)
 
     log.info(T('reading input paths'))
-    # tee is used to keep memory footprint low
-    # though it would be faster to create a list from path_bundle
-    #_, pdbs = it.tee(libio.read_path_bundle(pdbs, ext='.pdb'))
-    pdbs_ = sorted(libio.read_path_bundle(pdbs, ext='.pdb'), key=lambda x: x.stem)
-    #log.info(S('done'))
-    #pdbids = libpdb.PDBList(pdbs_)
-    #print(len(pdbids))
-    #print(len(pdbs_))
+    #pdbs_ = sorted(libio.read_path_bundle(pdbs, ext='.pdb'), key=lambda x: x.stem)
+    pdbs2operate = FileReaderIterator(pdb_files, ext='.pdb')
 
     manager = Manager()
     mdict = manager.dict()
 
-    libmulticore.pool_function(
+    pool_function(
         get_fastas,
-        pdbs_,
+        pdbs2operate,
         mdict=mdict,
         ncores=ncores,
         )
 
-    libio.write_text('\n'.join(f'{k}|{v}' for k,v in mdict.items()), output)
-
-
-def get_fastas(pdbfile, mdict=None):
-    structure = libstructure.Structure(pdbfile)
-    structure.build()
-    fasta = structure.fasta
-    assert len(fasta) == 1
-    mdict[libpdb.PDBIDFactory(pdbfile)] = next(iter(fasta.values()))
-    #out_data.append('{}|{}'.format(pdbid, next(iter(fasta.values()))))
-
-    return
+    #libio.write_text('\n'.join(f'{k}|{v}' for k,v in mdict.items()), output)
+    save_dictionary(mdict, output)
 
 
 if __name__ == '__main__':
-    maincli()
+    libcli.maincli(ap, main)
